requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=2.6.0
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv>=1.0.0
